fix: Polkadot-API (PAPI) runtime API discovery
